Open CV Libraries
For anyone who has been having trouble running the application, the open
cv libraries have been uploaded. You must link the libraries yourself.
Remove the existing link to open cv (this can be done in eclipse) or
create a new project with just the source files. Sorry if this is too
long winded. If you are using eclipse use this link to see how to do it:

http://docs.opencv.org/doc/tutorials/introduction/java_eclipse/java_eclipse.html